(PC-25822)[EAC] feat: rm favorite api module (move code to favorites)
Two similar educational api modules existed: favorite and favorites. It
does not make sense. Code from favorite has been moved to favorites.
