+++ api/src/pcapi/core/educational/api/favorites.py
@@ -4,6 +4,7 @@ import sqlalchemy as sa
 
 from pcapi.core.educational import models as educational_models
 from pcapi.core.offerers import models as offerers_models
+from pcapi.models import db
 from pcapi.repository import repository
 
 
@@ -128,3 +129,19 @@ def _offer_template_favorites_count(redactor_id: int) -> int:
     ]
 
     return len(favorite_offer_templates)
+
+
+def is_offer_a_redactor_favorite(offer_id: int, redactor_id: int) -> bool:
+    query = educational_models.CollectiveOfferEducationalRedactor.query.filter_by(
+        collectiveOfferId=offer_id,
+        educationalRedactorId=redactor_id,
+    )
+    return db.session.query(query.exists()).scalar()
+
+
+def is_offer_template_a_redactor_favorite(offer_id: int, redactor_id: int) -> bool:
+    query = educational_models.CollectiveOfferTemplateEducationalRedactor.query.filter_by(
+        collectiveOfferTemplateId=offer_id,
+        educationalRedactorId=redactor_id,
+    )
+    return db.session.query(query.exists()).scalar()
